+++ internal/query/sign.gen.go
@@ -172,12 +172,12 @@ type ISignDo interface {
 	UnderlyingDB() *gorm.DB
 	schema.Tabler
 
-	GetByID(id int) (result *model.Sign, err error)
+	GetByID(id uint64) (result *model.Sign, err error)
 }
 
 // GetByID
 // SELECT * FROM @@table WHERE id=@id
-func (s signDo) GetByID(id int) (result *model.Sign, err error) {
+func (s signDo) GetByID(id uint64) (result *model.Sign, err error) {
 	var params []interface{}
 
 	var generateSQL strings.Builder
